In particular, this online insurance management system project in Python Django focuses mainly on dealing with insurance policies.
The system also shows all of the possible categories, along with their policies. Additionally, the system enables for the management of client records. This project is clearly separated into two sections: Customer and Admin Panel. In this online application's summary, a consumer may easily register and begin utilising it. Initially, the system displays all of the student's possible insurance coverage. He/she may, in fact, just apply for it. Each application, however, necessitates action from the administrator, which is contingent on their acceptance or disapproval. Every application will be marked as pending by default. In addition, the customer may check policy history and submit a variety of questions, as well as read the admin's answer.